# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API Laravel",
        "description": "Implémentation de Swagger avec L5-Swagger.",
        "contact": {
            "name": "KANHONOU AUSCENCE",
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local de l'API"
        }
    ],
    "paths": {
...
        "/api/projet": {
            "get": {
                "tags": [
                    "Projets"
                ],
                "summary": "Récupère la liste des projets",
                "description": "Retourne une liste paginée de projets, incluant les relations 'users' et 'organisme'.",
                "operationId": "listProjets",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page pour la pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des projets récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projets"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projet/promuleguer": {
            "get": {
                "tags": [
                    "Projets"
                ],
                "summary": "Récupère les projets promulgués",
                "description": "Retourne la liste des projets ayant le statut promulgué.",
                "operationId": "listProjetPromuleguer",
                "responses": {
                    "200": {
                        "description": "Liste des projets promulgués",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projets"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projet/nonPromulegue": {
            "get": {
                "tags": [
                    "Projets"
                ],
                "summary": "Récupère les projets non promulgués",
                "description": "Retourne la liste des projets n'ayant pas le statut promulgué.",
                "operationId": "listProjetNonPromulegue",
                "responses": {
                    "200": {
                        "description": "Liste des projets non promulgués",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projets"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projet/avoter": {
            "get": {
                "tags": [
                    "Projets"
                ],
                "summary": "Récupère les projets à voter",
                "description": "Retourne la liste des projets en attente de vote.",
                "operationId": "listProjetAvoter",
                "responses": {
                    "200": {
                        "description": "Liste des projets à voter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projets"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projet/voter/{projet_id}": {
            "post": {
                "tags": [
                    "Projets"
                ],
                "summary": "Vote pour ou contre un projet",
                "description": "Enregistre le vote d'un utilisateur sur un projet de lois.",
                "operationId": "voterProjet",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/projet--id"
                    }
                ],
                "requestBody": {
                    "description": "Données du vote",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vote"
                                ],
                                "properties": {
                                    "vote": {
                                        "description": "Vote de l'utilisateur (true=pour, false=contre)",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "commentaire": {
                                        "description": "Commentaire optionnel du vote",
                                        "type": "string",
                                        "example": "En accord avec le projet"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vote enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vote enregistré avec succès"
                                        },
                                        "vote": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "projet_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "vote": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "commentaire": {
                                                    "type": "string",
                                                    "example": "En accord avec le projet"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données de vote invalides"
                    },
                    "404": {
                        "description": "Projet non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projet/store": {
            "post": {
                "tags": [
                    "Projets"
                ],
                "summary": "Crée un nouvel projet",
                "description": "Crée et enregistre un nouveau projet de lois en base de données.",
                "operationId": "storeProjet",
                "requestBody": {
                    "description": "Données du projet à créer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Projet"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projet créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projet"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Les données envoyées sont incorrectes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Projet créé avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Projet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projet/show/{projet_id}": {
            "get": {
                "tags": [
                    "Projets"
                ],
                "summary": "Affiche les détails d'un projet",
                "description": "Retourne les informations détaillées d'un projet spécifique, incluant ses utilisateurs et son organisme.",
                "operationId": "showProjet",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/projet--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du projet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projet non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projet/update/{projet_id}": {
            "put": {
                "tags": [
                    "Projets"
                ],
                "summary": "Met à jour un projet",
                "description": "Met à jour les informations d'un projet existant.",
                "operationId": "updateProjet",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/projet--id"
                    }
                ],
                "requestBody": {
                    "description": "Données du projet à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Projet"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Projet mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projet"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Les données envoyées sont incorrectes"
                    },
                    "404": {
                        "description": "Projet non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projet/delete/{projet_id}": {
            "delete": {
                "tags": [
                    "Projets"
                ],
                "summary": "Supprime un projet",
                "description": "Supprime un projet de lois de la base de données.",
                "operationId": "destroyProjet",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/projet--id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Projet supprimé avec succès"
                    },
                    "404": {
                        "description": "Projet non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Projet": {
                "title": "Projet",
                "required": [
                    "title",
                    "filePath"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identifiant du projet de lois",
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Titre du projet de lois",
                        "type": "string",
                        "example": "Loi de finances 2025"
                    },
                    "filePath": {
                        "title": "filePath",
                        "description": "Chemin du fichier du projet de lois",
                        "type": "string",
                        "example": "/files/projet_1.pdf"
                    },
                    "etat": {
                        "title": "etat",
                        "description": "État du projet (en_cours, promulgue, etc)",
                        "type": "string",
                        "example": "en_cours"
                    },
                    "avoter": {
                        "title": "avoter",
                        "description": "Indique si le projet est à voter",
                        "type": "boolean",
                        "example": true
                    },
                    "cloturevoter": {
                        "title": "cloturevoter",
                        "description": "Date de clôture du vote",
                        "type": "string",
                        "example": "2025-12-31T23:59:59Z",
                        "nullable": true
                    },
                    "organisme_id": {
                        "title": "organisme_id",
                        "description": "Identifiant de l'organisme responsable",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-01-02T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Projets": {
                "title": "Projets",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Projet"
                        }
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
...
            "projet--id": {
                "name": "projet_id",
                "in": "path",
                "description": "Identifiant de projet",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}